fix: jpa repo di 오류 수정

+++ src/main/kotlin/com/wheretopop/infrastructure/popup/external/popply/PopupPopplyProcessor.kt
@@ -38,7 +38,7 @@ class PopupPopplyProcessor(
     override fun save(popupDetail: PopupDetail, popupId: PopupId) {
         val popupPopplyEntity = PopupPopplyEntity.of(
             popupDetail = popupDetail,
-            popupId = popupId.toLong(),
+            popupId = popupId,
         )
         popupPopplyRepository.save(popupPopplyEntity)
     }
@@ -155,7 +155,7 @@ class PopupPopplyProcessor(
 
         val popupPopplyEntity = PopupPopplyEntity.of(
             popupDetail = popupDetailData,
-            popupId = generatedPopupId.toLong(),
+            popupId = generatedPopupId,
         )
 
         popupPopplyRepository.save(popupPopplyEntity)
